client/tools/png2c/requirements.txt
Pillow
argparse

